feat: update pi test code

+++ pi/main.go
@@ -13,11 +13,11 @@ func main() {
 
 	x := 1.0
 	pi := 1.0
-	stop := float64(rounds + 2)
+	stop := rounds + 2
 
-	for i := 2.0; i <= stop; i++ {
+	for i := 2; i <= stop; i++ {
 		x = -x
-		pi += x / (2.0*i - 1.0)
+		pi += x / float64(2*i-1)
 	}
 
 	pi *= 4.0
